Add basic support for migration.
