+++ caravel/testing/integration_test.py
@@ -1,6 +1,6 @@
 import urlparse, os
 from caravel import app
-from caravel.daemons import migration
+from caravel.daemons import replication
 from caravel.storage import helpers
 from flask import jsonify
 
@@ -18,7 +18,7 @@ def run_integration_test():
         return open(path)
 
     # Migrate a listing from the old site.
-    migration.pull_from_listing("listing-a", _urlopen=opener)
+    replication.pull_from_old_marketplace("listing-a", _urlopen=opener)
 
     # Verify that the listing was created.
     data = helpers.lookup_listing("listing-a")
@@ -31,7 +31,5 @@ def run_integration_test():
         data["seller"] != "seller@uchicago.edu"):
 
         raise ValueError("migration failed")
-
-    print "Everything worked!"
 
     return "ok"
